backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
google-genai==1.0.0
PyPDF2==3.0.1
python-docx==1.1.2